tools/content_manager/requirements.txt
google-auth
requests
python-dotenv
PyYAML
ruamel.yaml
pydantic~=2.6.0
click
